requirements.txt
discord.py
python-dotenv
fastf1
schedule
pymongo